Adapting interface PR to main branch changes
