+++ py/PyBTLS/py/utils_interface/_vehicle_event_writer.py
