--- py_src/PyBTLS/py_module/default_files/constant_vehicle.csv
@@ -1,6 +1,0 @@
-0.0,0.0
-3.0,58.5
-5.0,135
-1.1,85.5
-1.1,85.5
-0.0,85.5
