--- py_src/PyBTLS/py_module/default_files/kernels.csv
@@ -1,3 +1,0 @@
-1.0,0.08
-1.0,0.05
-1.0,0.02
