+++ tests/PyBTLS_interface/Input/IL.txt
@@ -1,0 +1,17 @@
+3
+1,4
+0.0,0.0
+37.49,9.3725
+37.730000000000004,9.202499999999997
+50.0,0.0
+2,4
+0.0,0.0
+12.49,9.3675
+12.76,9.31
+50.0,0.0
+3,5
+0.0,0.0
+12.48,-0.24960000000000002
+12.49,-0.24980000000000002
+12.5,0.75
+50.0,0.0
