+++ tests/PyBTLS_interface/Input/Traffic/ASall.csv
@@ -1,0 +1,15 @@
+1.000,55.3,5.1,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+
+0.143,32.8,6.3,1.000,12.2,1.5,0.000,0.0,0.0,0.000,0.0,0.0,
+0.252,44.3,4.5,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+0.605,53.9,2.3,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+
+1.000,32.2,1.6,0.130,46.5,8.4,0.806,13.0,1.3,0.000,0.0,0.0,
+0.000,0.0,0.0,0.616,62.3,4.9,0.194,19.4,1.3,0.000,0.0,0.0,
+0.000,0.0,0.0,0.254,61.8,2.0,0.000,0.0,0.0,0.000,0.0,0.0,
+
+1.000,32.4,1.7,0.041,42.4,3.0,1.000,12.0,1.7,1.000,12.1,1.5,
+0.000,0.0,0.0,0.959,54.3,3.4,0.000,0.0,0.0,0.000,0.0,0.0,
+0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
