+++ tests/PyBTLS_interface/Input/Traffic/ATW.csv
@@ -1,0 +1,15 @@
+0.109,188,27.3,0.109,188,27.3,0,0,0,0,0,0,0,0,0
+0.847,193.9,6.4,0.847,193.9,6.4,0,0,0,0,0,0,0,0,0
+0.045,167.6,6.1,0.045,167.6,6.1,0,0,0,0,0,0,0,0,0
+,,,,,,,,,,,,,,
+0.109,188,27.3,0,0,0,0,0,0,0,0,0,0,0,0
+0.847,193.9,6.4,0,0,0,0,0,0,0,0,0,0,0,0
+0.045,167.6,6.1,0,0,0,0,0,0,0,0,0,0,0,0
+,,,,,,,,,,,,,,
+1,196,0,1,184,0,1,199,0,1,199,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+,,,,,,,,,,,,,,
+1,196,0,1,184,0,1,199,0,1,199,0,1,199,0
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
