+++ tests/PyBTLS_interface/Input/Traffic/AW2&3.csv
@@ -1,0 +1,11 @@
+0.316,30.9,2.5,0.317,69.1,2.5,0.000,0.0,0.0,
+0.684,33.9,3.9,0.683,66.1,3.9,0.000,0.0,0.0,
+0.000,0.0,0.0,0.000,0.0,0.0,0.000,0.0,0.0,
+
+0.515,26.6,2.1,0.132,31.8,5.7,0.715,23.2,2.3,
+0.459,28.3,6.3,0.233,39.5,4.8,0.200,34.1,3.0,
+0.026,21.5,43.3,0.635,50.3,3.0,0.085,40.8,2.3,
+
+ 
+ 
+
