+++ tests/PyBTLS_interface/Input/Traffic/AW4&5.csv
@@ -1,0 +1,25 @@
+24.9,29.5,45.6,3.9,3.2,3.9,
+25.1,36.2,38.7,5.0,7.6,7.8,
+20.8,36.8,42.4,5.4,7.4,7.4,
+20.6,36.8,42.6,4.9,5.2,7.2,
+21.4,37.3,41.3,3.7,4.4,5.2,
+19.8,36.9,43.4,3.3,4.3,5.4,
+17.7,35.4,46.9,2.8,3.9,5.3,
+15.8,34.4,49.7,2.2,3.2,4.2,
+14.8,34.3,50.9,1.8,2.6,3.3,
+14.6,34.1,51.3,1.7,2.5,3.0,
+14.4,32.7,52.9,2.1,2.5,3.4,
+13.9,31.9,54.1,0.7,2.0,2.4,
+
+0.0,0.0,0.0,0.0,0.0,0.0,
+21.0,26.7,52.3,4.1,5.0,6.3,
+20.8,30.0,49.2,4.3,8.6,8.1,
+17.2,35.3,47.4,3.8,6.6,7.6,
+19.3,34.7,46.1,4.7,5.0,7.4,
+19.5,34.6,45.9,3.8,4.2,5.9,
+17.3,34.3,48.5,2.8,4.3,5.6,
+15.3,33.2,51.5,2.2,4.2,5.2,
+14.3,32.2,53.6,1.7,3.7,4.5,
+13.7,31.2,55.1,1.5,3.3,4.0,
+13.5,29.9,56.6,1.4,2.9,3.5,
+13.4,28.8,57.8,1.6,2.8,3.6,
