+++ tests/PyBTLS_interface/Input/Traffic/CLASS%.csv
@@ -1,0 +1,4 @@
+23.0,21.9,
+2.8,2.3,
+31.7,31.0,
+42.5,44.8,
