+++ tests/PyBTLS_interface/Input/Traffic/FlowR.csv
@@ -1,0 +1,24 @@
+153.8,92.2,
+131,79.6,
+131.8,67,
+123.8,74.8,
+114,81.6,
+121.2,94.8,
+141.2,102.4,
+155.4,121.2,
+154,127.4,
+141,127.2,
+126.4,112.2,
+101.6,111.4,
+95.8,110.2,
+88.2,146,
+93,160.4,
+109,152,
+124.2,151,
+151,167.2,
+141.8,179.6,
+172.2,164.8,
+141.4,206.6,
+148,228.4,
+157.2,189,
+159.4,138.8,
