+++ tests/PyBTLS_interface/Input/Traffic/GVWpdf.csv
@@ -1,0 +1,7 @@
+1.000,248.0,10.0,0.064,43.8,2.3,0.060,85.2,23.2,0.476,261.6,53.6,0.160,296.2,55.1,
+0.000,0.0,0.0,0.642,81.3,67.3,0.368,223.0,23.6,0.524,386.6,45.2,0.834,412.2,55.1,
+0.000,0.0,0.0,0.294,163.0,26.2,0.572,220.7,51.2,0.000,386.6,45.2,0.006,604.3,56.4,
+
+1.000,222.0,10.0,0.177,41.5,5.5,0.094,99.1,28.1,0.037,265.4,140.2,0.149,293.8,59.6,
+0.000,0.0,0.0,0.456,104.8,45.0,0.711,241.0,57.5,0.450,284.3,57.0,0.839,456.8,60.8,
+0.000,0.0,0.0,0.367,188.0,31.4,0.195,252.7,16.4,0.513,425.4,48.9,0.012,583.1,25.2,
