+++ tests/PyBTLS_interface/Input/Traffic/NHM.csv
@@ -1,0 +1,18 @@
+15,0,0,0
+0,0.011855673,-0.014268241,0.004048786
+0,0.039251526,-0.05978246,0.02212043
+70,-0.004412997,0.054824101,-0.066907905
+80,-0.004685721,0.052127816,-0.053475193
+90,0.001537014,0.020896587,-0.013787689
+100,-0.003853623,0.064555837,-0.069172155
+110,-0.002530238,0.054511802,-0.059714977
+120,-0.001307981,0.048010242,-0.051645258
+130,-0.000487752,0.049738587,-0.057875119
+140,-0.004995115,0.081041256,-0.086465967
+150,-0.004547469,0.080310658,-0.083351351
+160,-0.004938412,0.092219287,-0.105416601
+170,-0.005000644,0.086893379,-0.097048852
+180,0.001987438,0.052114614,-0.058245039
+190,0.003366332,0.044909211,-0.063187142
+210,0.000379907,0.068461437,-0.077769612
+230,-0.006466786,0.117770005,-0.141174818
