+++ tests/PyBTLS_interface/Input/bridge.txt
@@ -1,0 +1,9 @@
+1,50.0,1,4
+1,1,0.0
+1,1,1.0
+2,1,0.0
+2,1,1.0
+3,1,0.0
+2,2,1.0
+4,1,0.0
+2,3,1.0
