+++ tests/PyBTLS_interface/Input/lane_flow.csv
@@ -1,0 +1,25 @@
+1.0,1.0,,,,,,,
+0.0,153.8,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+1.0,131.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+2.0,131.8,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+3.0,123.8,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+4.0,114.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+5.0,121.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+6.0,141.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+7.0,155.4,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+8.0,154.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+9.0,141.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+10.0,126.4,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+11.0,101.6,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+12.0,95.8,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+13.0,88.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+14.0,93.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+15.0,109.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+16.0,124.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+17.0,151.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+18.0,141.8,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+19.0,172.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+20.0,141.4,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+21.0,148.0,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+22.0,157.2,248.0,10.0,80.0,23.0,2.8,31.7,42.5
+23.0,159.4,248.0,10.0,80.0,23.0,2.8,31.7,42.5
